+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Include new error hint message in API responses [#208](https://github.com/CartoDB/carto-react/pull/208/) 
 - Fix POST requests to SQL API not managing access_token properly [#210](https://github.com/CartoDB/carto-react/pull/210) 
 - Fix echarts update problems [#209](https://github.com/CartoDB/carto-react/pull/209)
+- Fix issue with TimeSeriesWidget markLine if xAxis is undefined [#211](https://github.com/CartoDB/carto-react/pull/211)
 
 ## 1.1.0-beta.1 (2021-10-14)